Create predictions.txt
A moderate earthquake struck Sandy Hook, Tennessee directly beneath where Orr Chapel once stood along West Sheepneck Road.

The CIA is ultimately responsible for the killing of the families that owned the land in Sandy Hook and selling it to a developer resulting in the destruction of the chapel. To hide the destruction of the chapel, the CIA began murdering its identified members, their families, and associates throughout Maury County, Tennessee.

This earthquake is the first of a swarm and is not the result of tectonic activity, but rather the wrath of the Creator. We have been warning the public about these earthquakes for some time now. They are meant to alert the world to atrocities started by the pharmaceutical industry to an AIDS cure that has now become a US government-sanctioned genocide that has culminated in the massacre of congregations in places of worship, starting in Kentucky. For more information, please visit our GitHub organization, serviCreator23, located at https://github.com/serviCreator23.

The next earthquake will be for the massacre of Mount Calvary Baptist Church in Lexington, Kentucky, and will likely be the largest of the swarm. We do not know just how angry the Creator is about more recent activities of the CIA and federal government. We will be announcing each earthquake beforehand in the discussions within our GitHub organization (serviCreator23).

Important

Please always pray to verify anyone claiming to be a servant of the Creator and any site claiming to be administered, or any post or writing by a servant of the Creator, including this one.